+++ tfhe/src/integer/server_key/radix_parallel/tests_unsigned/test_slice.rs
@@ -60,7 +60,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks);
 
@@ -104,7 +104,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks);
 
@@ -151,7 +151,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks);
 
@@ -195,7 +195,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks);
 
@@ -239,7 +239,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks.clone());
 
@@ -289,7 +289,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks.clone());
 
@@ -339,7 +339,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks.clone());
 
@@ -389,7 +389,7 @@ where
     let mut rng = rand::thread_rng();
 
     // message_modulus^vec_length
-    let modulus = param.message_modulus().0.pow(NB_CTXT as u32) as u64;
+    let modulus = param.message_modulus().0.pow(NB_CTXT as u32);
 
     executor.setup(&cks, sks.clone());
